+++ NEWS.md
@@ -1,6 +1,18 @@
 > Copyright © 2016 RTE Reseau de transport d’electricite
 
-# antaresEditObject 0.9.2
+
+# antaresEditObject 0.9.3.9000
+(cf. Antares v9.3 changelog)
+
+NEW FEATURES :  
+
+* `createClusterST()`/`editClusterST()` :  
+  - **New properties** (*allow-overflow*, see list of properties according to study version of Antares with `storage_values_default()`) 
+* `createClusterST()`/`editClusterST()` :  
+  - **New dimension of time series** (all matrices will be of (8760, N), noting that N >= 1)
+  - **Suppression of properties** (refreshtimeseries,refreshintervalload, refreshintervalhydro, refreshintervalwind, refreshintervalthermal, refreshintervalsolar from settings/generaldata.ini)
+  
+# antaresEditObject 0.9.2.9000
 (cf. Antares v9.2 changelog)
 
 NEW FEATURES :